[Fix] 데미지 이펙트 소수점 제거
1. 받아온 데미지를 int로 형변환 후 이펙트 설정

+++ Assets/WorkSpace/BSM/Scripts/Damage/DamageEffect.cs
@@ -51,7 +51,7 @@ public class DamageEffect : MonoBehaviour
         
         //피해 입은 데미지 설정
         DamageText damageText = instance.GetComponent<DamageText>();
-        damageText.SetDamageText(damage.ToString(), unitType);
+        damageText.SetDamageText(((int)damage).ToString(), unitType);
         
         instance.SetActive(true);
     }
